refactor!: move fetch into seperate package (#1913)
