+++ packages/protocol-fetch/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@libp2p/autonat",
+  "name": "@libp2p/fetch",
   "version": "1.0.0",
-  "description": "Implementation of Autonat Protocol",
+  "description": "Implementation of the Fetch Protocol",
   "license": "Apache-2.0 OR MIT",
-  "homepage": "https://github.com/libp2p/js-libp2p/tree/master/packages/protocol-autonat#readme",
+  "homepage": "https://github.com/libp2p/js-libp2p/tree/master/packages/protocol-fetch#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/libp2p/js-libp2p.git"
@@ -45,26 +45,28 @@
     "dep-check": "aegir dep-check"
   },
   "dependencies": {
-    "@libp2p/interface": "^0.1.2",
-    "@libp2p/interface-internal": "^0.1.5",
-    "@libp2p/peer-id": "^3.0.2",
-    "@libp2p/peer-id-factory": "^3.0.4",
-    "@libp2p/logger": "^3.0.2",
-    "@multiformats/multiaddr": "^12.1.5",
-    "it-first": "^3.0.1",
-    "it-length-prefixed": "^9.0.1",
-    "it-map": "^3.0.3",
-    "it-parallel": "^3.0.0",
+    "@libp2p/interface": "^0.1.4",
+    "@libp2p/interface-internal": "^0.1.7",
+    "@libp2p/logger": "^3.0.4",
+    "it-first": "^3.0.3",
+    "it-length-prefixed": "^9.0.3",
     "it-pipe": "^3.0.1",
-    "private-ip": "^3.0.0",
-    "protons-runtime": "^5.0.0",
-    "uint8arraylist": "^2.4.3"
+    "protons-runtime": "^5.2.0",
+    "uint8arraylist": "^2.4.3",
+    "uint8arrays": "^4.0.6"
   },
   "devDependencies": {
-    "aegir": "^41.0.2",
-    "it-all": "^3.0.1",
-    "it-pushable": "^3.2.0",
-    "sinon": "^17.0.0",
-    "sinon-ts": "^1.0.0"
+    "@chainsafe/libp2p-noise": "^13.0.2",
+    "@chainsafe/libp2p-yamux": "^5.0.0",
+    "@libp2p/interface-compliance-tests": "^4.1.2",
+    "@libp2p/peer-id-factory": "^3.0.6",
+    "@libp2p/peer-store": "^9.0.7",
+    "@libp2p/tcp": "^8.0.10",
+    "aegir": "^40.0.8",
+    "datastore-core": "^9.2.3",
+    "delay": "^6.0.0",
+    "libp2p": "^0.46.16",
+    "sinon": "^17.0.1",
+    "sinon-ts": "^2.0.0"
   }
 }
